requirements.txt
asgiref==3.7.2
authorizenet==1.1.4
Django==4.2.2
django-ckeditor==6.5.0
django-environ==0.8.1
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-oauth-toolkit==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
Pillow==9.2.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
squareup==29.0.0.20230720
typing_extensions==4.7.1
tzdata==2023.3
oauthlib==3.2.2
